requirements.txt
flask
werkzeug
psycopg2-binary
pytest
